backend/requirements.txt
flask
flask-cors
python-dotenv
pymupdf
pdfplumber
openai
